third_party/WebKit/LayoutTests/inspector/sources/debugger/async-callstack-events-expected.txt

Tests asynchronous call stacks for various DOM events.

Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) onHashChange (async-callstack-events.html:51)
    [hashchange]
    0) doTestHashChange (async-callstack-events.html:45)
    1) testFunction (async-callstack-events.html:11)

Call stack:
    0) onSelectionChange (async-callstack-events.html:39)
    [selectionchange]
    0) setSelection (async-callstack-events.html:24)
    1) doTestSelectionChange (async-callstack-events.html:31)
    2) testFunction (async-callstack-events.html:11)
